Finishing touches

+++ src/metrics.h
@@ -1,6 +1,5 @@
 #pragma once
-#include <math.h>
 #include "nodes.h"
 
-double get_distance(const Node *from, const Node *to);
-double get_heuristic(const Node *from, const Node *to);
+float get_distance(const Node *from, const Node *to);
+float get_heuristic(const Node *from, const Node *to);
